Split Pico specifict utilities to separate util_rp2040.c
